flats_scrapping/requirements.txt
pyopenssl==22.0.0
cryptography==38.0.4
scrapy==2.5.1
psycopg2-binary==2.9.1
